+++ OOS/OOS.Domain/Contacts/Models/Feedback.cs
@@ -1,5 +1,4 @@
-﻿using MongoDB.Bson.Serialization.Attributes;
-using OOS.Infrastructure.Domain;
+﻿using OOS.Infrastructure.Domain;
 using OOS.Infrastructure.Mongodb;
 using System;
 using System.Collections.Generic;
@@ -7,9 +6,8 @@ using System.Text;
 
 namespace OOS.Domain.Contacts.Models
 {
-    public class Email : AuditableEntityBase, IAggregateRoot
+    public class Feedback : AuditableEntityBase, IAggregateRoot
     {
-
         public string Id { set; get; }
 
         public string ToEmail { set; get; }
@@ -18,5 +16,6 @@ namespace OOS.Domain.Contacts.Models
 
         public string Content { set; get; }
 
+        public Boolean Status  { set; get; }
     }
 }
