+++ OOS/OOS.Presentation.ApplicationLogic/Contacts/EmailBusinessLogicAutoMapper.cs
@@ -1,6 +1,7 @@
 ﻿using AutoMapper;
 using OOS.Domain.Contacts.Models;
 using OOS.Presentation.ApplicationLogic.Contacts.Messages;
+using OOS.Presentation.ApplicationLogic.CustomerFeedback;
 using System;
 using System.Collections.Generic;
 using System.Text;
@@ -12,7 +13,9 @@ namespace OOS.Presentation.ApplicationLogic.Contacts
         public EmailBusinessLogicAutoMapper()
         {
             CreateMap<CreateEmailSubscribeRequest, EmailSubsribe>();
-            CreateMap<SentEmailRequest, Email>();
+            CreateMap<SentEmailRequest, Feedback>();
+            CreateMap<EditFeedBackRequest, Feedback>();
+            CreateMap<Feedback, EditFeedBackResponse>();
         }
     }
 }
