+++ OOS/OOS.Presentation.ApplicationLogic/Contacts/IEmailBusinessLogic.cs
@@ -1,5 +1,6 @@
 ﻿using OOS.Domain.Contacts.Models;
 using OOS.Presentation.ApplicationLogic.Contacts.Messages;
+using OOS.Presentation.ApplicationLogic.CustomerFeedback;
 using System;
 using System.Collections.Generic;
 using System.Text;
@@ -16,6 +17,16 @@ namespace OOS.Presentation.ApplicationLogic.Contacts
         CreateEmailSubscribeResponse CreateEmailSubscribe(CreateEmailSubscribeRequest request);
 
         void DeleteEmailSubsribe(string id);
+
+        Feedback GetFeedBack(string id);
+
+        List<Feedback> GetEmailFeedBack();
+
+        void DeleteFeedback(string id);
+
+        SentEmailResponse CreateFeedBack(SentEmailRequest request);
+
+        EditFeedBackResponse EditFeedBack(string id, EditFeedBackRequest request);
 
     }
 }
