+++ OOS/OOS.Presentation.ApplicationLogic/Contacts/Messages/SentEmailRequest.cs
@@ -7,10 +7,14 @@ namespace OOS.Presentation.ApplicationLogic.Contacts.Messages
 {
     public class SentEmailRequest: RequestBase
     {
+        public string Id { set; get; }
+
         public string ToEmail { set; get; }
 
         public string Subject { set; get; }
 
         public string Content { set; get; }
+
+        public Boolean Status = (false);
     }
 }
