--- OOS/OOS.Presentation.ApplicationLogic/CustomerFeedback/Messages/CreateFeedbackResponse.cs
@@ -1,12 +1,0 @@
-﻿using OOS.Presentation.ApplicationLogic.Common;
-using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace OOS.Presentation.ApplicationLogic.Contacts.Messages
-{
-    public class CreateFeedbackResponse : ResponseBase
-    {
-
-    }
-}
