+++ OOS/OOS.Presentation.WebAPIs/Controllers/ContactController.cs
@@ -5,6 +5,8 @@ using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
 using OOS.Presentation.ApplicationLogic.Contacts;
 using OOS.Presentation.ApplicationLogic.Contacts.Messages;
+using OOS.Presentation.ApplicationLogic.CustomerFeedback;
+
 
 // For more information on enabling Web API for empty projects, visit https://go.microsoft.com/fwlink/?LinkID=397860
 
@@ -33,6 +35,7 @@ namespace OOS.Presentation.WebAPIs.Controllers
         public IActionResult Post([FromBody] SentEmailRequest value)
         {
             _emailsBusinessLogic.SentEmail(value);
+            _emailsBusinessLogic.CreateFeedBack(value);
             return Ok();
         }
 
@@ -40,7 +43,7 @@ namespace OOS.Presentation.WebAPIs.Controllers
         [HttpPost]
         public IActionResult EmailSubsribe([FromBody] CreateEmailSubscribeRequest value)
         {
-            _emailsBusinessLogic.CreateEmailSubscribe(value);
+            _emailsBusinessLogic.CreateEmailSubscribe(value); 
             return Ok();
         }
 
@@ -49,6 +52,39 @@ namespace OOS.Presentation.WebAPIs.Controllers
         {
             _emailsBusinessLogic.DeleteEmailSubsribe(id);
             return Ok();
+        }
+
+        [Route("getFeedback")]
+        [HttpGet]
+        public IActionResult GetFeedBack()
+        {
+            var listEmailFeedBack = _emailsBusinessLogic.GetEmailFeedBack();
+            return Ok(listEmailFeedBack);
+        }
+
+        // GET api/<controller>/5
+        [Route("getFeedback/{id}")]
+        [HttpGet]
+        public IActionResult Get(string id)
+        {
+            var result = _emailsBusinessLogic.GetFeedBack(id);
+            return Ok(result);
+        }
+
+        [Route("getFeedback/{id}")]
+        [HttpDelete]
+        public IActionResult DeleteFeedback(string id)
+        {
+            _emailsBusinessLogic.DeleteFeedback(id);
+            return Ok();
+        }
+
+        [Route("getFeedback/{id}")]
+        [HttpPut]
+        public IActionResult PutFeedBack(string id, [FromBody] EditFeedBackRequest request)
+        {
+            var rs = _emailsBusinessLogic.EditFeedBack(id, request);
+            return Ok(rs);
         }
 
     }
